Rename fcoin/endpoint => fcoin/api/endpoint
